+++ src/main/java/me/Fupery/ArtMap/IO/Database/Map.java
@@ -1,20 +1,21 @@
 package me.Fupery.ArtMap.IO.Database;
 
-import me.Fupery.ArtMap.ArtMap;
-import me.Fupery.ArtMap.IO.CompressedMap;
-import me.Fupery.ArtMap.IO.ErrorLogger;
-import me.Fupery.ArtMap.Painting.GenericMapRenderer;
-import me.Fupery.ArtMap.Utils.Reflection;
+import java.io.File;
+import java.io.FileInputStream;
+import java.io.IOException;
+import java.util.Arrays;
+
 import org.bukkit.Bukkit;
 import org.bukkit.World;
 import org.bukkit.entity.Player;
 import org.bukkit.map.MapRenderer;
 import org.bukkit.map.MapView;
 
-import java.io.File;
-import java.io.FileInputStream;
-import java.io.IOException;
-import java.util.Arrays;
+import me.Fupery.ArtMap.ArtMap;
+import me.Fupery.ArtMap.IO.CompressedMap;
+import me.Fupery.ArtMap.IO.ErrorLogger;
+import me.Fupery.ArtMap.Painting.GenericMapRenderer;
+import me.Fupery.ArtMap.Utils.Reflection;
 
 public class Map {
 
@@ -29,7 +30,7 @@ public class Map {
     }
 
     public Map(MapView mapView) {
-        this.mapId = mapView.getId();
+		this.mapId = (short) mapView.getId();
         this.mapView = mapView;
     }
